requirements.txt
streamlit
yfinance
numpy
pandas
matplotlib
scikit-learn
tensorflow-cpu